requirements.txt
fastapi==0.88.0
pydantic==1.10.2
uvicorn==0.20.0
requests>=2.28.1
pymongo>=4.3.3
python-dotenv>=0.21.0
Routes~=2.5.1
pytz >= 2022.7